+++ NEWS.md
@@ -1,6 +1,7 @@
 # slfhelper 0.7.0
 
 * `ep_file_vars` and `indiv_file_vars` updated with the latest variable names. These now also return a character vector instead of a list.
+* `read_slf_` now returns a [tibble][tibble::tibble-package] 
 
 # slfhelper 0.6.0